refactoring code again
