+++ client/src/addCard/Dashboard.tsx
@@ -1,13 +1,8 @@
-import Status from '../dashboardComponents/Status';
+import Status from '../components/Status';
 import Cards from './Cards';
 import New from './New';
 import { useState } from 'react';
 import { ICardOrdered } from '../domain/useCase/orderCard';
-
-export interface Props{
-    dataCards: ICardOrdered[]
-    buttonIndex: number
-}
 
 const Dashboard = () => { 
     const [orderCards, setOrderCards ] = useState<ICardOrdered[]>([])
